Add comments to Actions yml file, more post WIP
Actions file is meant to be extra public-facing so it now should
serve as a sort of tutorial file.

Added a date to the "How my site works" post. Hopefully it ends up being accurate.
Also renamed it from "How this blog works" to "How my site works"

Committing the current state so that I can experiment with splitting it into two posts
instead of one. I think it's so long because I'm trying to satisfy two very different audiences,
and I don't really need to do that with one post.
We'll see!